- updated deposit/withdraw flow (checked usable TRY asset size instead of balance)

+++ src/main/java/com/brokerage/api/dto/DepositRequest.java
@@ -1,7 +1,6 @@
 package com.brokerage.api.dto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
-public record DepositRequest(UUID customerId, BigDecimal amount) {
+public record DepositRequest(UUID customerId, int amount) {
 }
