+++ src/main/java/com/brokerage/api/dto/WithdrawRequest.java
@@ -1,7 +1,6 @@
 package com.brokerage.api.dto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
-public record WithdrawRequest(UUID customerId, BigDecimal amount, String iban) {
+public record WithdrawRequest(UUID customerId, int amount, String iban) {
 }
